fix(#4843): Changes done over actions
