+++ deployability/modules/testing/tests/helpers/hostinformation.py
@@ -1,0 +1,50 @@
+import platform
+import os
+
+class HostInformation:
+    def __init__(self):
+        pass
+
+    def get_os_type(self):
+        """
+        It returns the os_type of host
+
+        Returns:
+            str: type of host (windows, linux, macos)
+        """
+        system = platform.system()
+
+        if system == 'Windows':
+            return 'windows'
+        elif system == 'Linux':
+            return 'linux'
+        elif system == 'Darwin':
+            return 'macos'
+        else:
+            return 'unknown'
+
+    def get_architecture(self):
+        """
+        It returns the arch of host
+
+        Returns:
+            str: arch (aarch64, x86_64, intel, apple)
+        """
+        return platform.machine()
+
+    def get_linux_distribution(self):
+        """
+        It returns the linux distribution of host
+
+        Returns:
+            str: linux distribution (deb, rpm)
+        """
+        if self.get_os_type() == 'linux':
+            package_managers = {
+                '/etc/debian_version': 'deb',
+                '/etc/redhat-release': 'rpm',
+            }
+
+            for file_path, package_manager in package_managers.items():
+                if os.path.exists(file_path):
+                    return package_manager
